MAINT: tweak script.

+++ build.ps1
@@ -18,6 +18,7 @@ $tensorrt_dir = "C:\local\TensorRT-7.0.0.11"
 $nvidia_codec_sdk_dir = "C:\local\Video_Codec_SDK_9.1.23"
 
 $update_vcpkg = $false
+$build_from_scratch = $false
 
 
 
@@ -44,6 +45,9 @@ if ($update_vcpkg) {
 
   # Install GLEW libraries.
   iex ".\vcpkg.exe install glew:x64-windows"
+
+  # Install GLFW libraries.
+  iex ".\vcpkg.exe install glfw3:x64-windows"
   echo `n
 }
 
@@ -61,16 +65,17 @@ echo `n
 
 
 echo "========================================================================="
-echo "Checking if directory ..\$build_dir exists"
-if (Test-Path ..\$build_dir) {
-  echo "Removing existing directory to rebuild from scratch..."
-  rm ..\$build_dir -r -fo
+if ($build_from_scratch) {
+  echo "Checking if directory ..\$build_dir exists"
+  if (Test-Path ..\$build_dir) {
+    echo "Removing existing directory to rebuild from scratch..."
+    rm ..\$build_dir -r -fo
+  }
+
+  echo "Creating directory `"..\$build_dir`" ..."
+  iex "New-Item -ItemType directory -Path ..\$build_dir"
+  echo "`n"
 }
-
-echo "Creating directory `"..\$build_dir`" ..."
-iex "New-Item -ItemType directory -Path ..\$build_dir"
-echo "`n"
-
 
 
 echo "========================================================================="
